pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "renamer"
version = "0.1.0"
description = "Smart File Renamer — 並べ替えと連番リネームのデスクトップアプリ"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "Pillow>=10.0.0",
]

[project.scripts]
renamer = "app.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.0.0",
    "pytest-qt>=4.4.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
strict = true
mypy_path = "src"
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
